gpu/docker-gpu/requirements.txt
tf2onnx
google-cloud-storage